Basic encryptor-decryptor.
You need python 3, cryptography.fernet and os libraries to run this program. The password to decrypt is "turk_kahvesi" by default. This is raw code. bot.py encrypts the files on the directory its in, while Decryptor.py decrypts. WARNING : The decryptor file should be in the same directory where the files are encrypted.